Add custom query encoding for prettyPrint, and fix a bug.
It turns out the prettyPrint flag is particular about how it's encoded (it
wants either lower-case true/false or 1/0), and apitools had been sending
values that got ignored.

Digging in, though, it turns out there was a subtler bug: we would never
include a query param whose value was False-y if it was in
StandardQueryParameters. This was just prettyPrint, but nice to fix anyway.