Restore chunked download support.
We lost support for chunked downloads in a merge; this PR restores chunked
download support, and adds tests. In addition, we add a `StreamMedia` method
to the `Download` class, following the `Upload` API. This function is
equivalent to `StreamInChunks` with an optional `ignore_chunksize` argument --
somehow `StreamInChunks(..., ignore_chunksize=True)` feels a bit awkward.

Relatedly, `Download.StreamInChunks` and `Download.GetRange` now have an
explicit `ignore_chunksize` argument.

I also updated the two integration tests, which really need to be updated to
run automatically.