Fix an issue with non-unique keys for custom JSON encoding.
Currently, JSON encoding uses `type.definition_name()` as the key for custom
JSON encodings. However, this makes use of the top-level `package` attribute,
which we purposely set to just the api name -- meaning we're guaranteed
conflicts in the case of simultaneous API versions for a given API.

The fix is straightforward -- we do a little extra work to pull out the
original module name, which we know will be unique.