Fix a bug in url-encoding of bytes fields.
Previously, we encoded url parameters based solely on *value* -- meaning that
we would try to normalize bytes fields to utf8, even when they should remain
bytes.

This just moves the logic up to encode while we're still holding on to type
information, so that we can check field types first.

(reviewed internally)