Fix : before / in paths for apitools.
Currently, if we have a top-level methodPath of the form `path:methodName`,
we'll pass this to `urlparse.urljoin`, where this is interpreted as
`scheme:netloc`. In cases other than media upload, this is **not** what we
want.

This fixes the issue by adding our own custom `urljoin`, which detects and
handles this case. We also add a test.

(This is basically a copy of
https://www.google.com/url?sa=D&q=https%3A%2F%2Fgithub.com%2Fgoogle%2Fgoogle-api-python-client%2Fcommit%2F7ee535d14007b3b05ca9cab1a35e93fb01657537
tweaked for apitools.)

(reviewed internally)