Fix a bug in handling for repeated any values.
If an API had a repeated any value marked as an additionalProperty, we'd fail
to decode, because we can't treat a JsonArray or a JsonValue with an
array_value as a list in Python.

We work around this in a painful way; in particular, there's a nasty inline
import because encoding.py needs to be split in two.

(reviewed internally)