Pin version of apitools for >=0.5 keep old behavior for earlier versions.
In order to keep compatibility with client packages already out there we keep
the behavior for clients built with apitools having version 0.4.N.
If apitools is version 0.4.N the clients will have version 0.4.REVISION_DATE and
apitools will be a requirement with >=0.4.8, <0.5.
If apitools is version >= 0.5 the clients will have version 0.5.N.REVISION_DATE
and apitools will be a requirement with 0.5.N (pinned version).