Cleanup some dead test code.
It turns out an ancient refactor (merging the gsutil fork) turned some tests
into dead code; I noticed that something was awry when the GCE header name was
updated but tests didn't fail.

The fix was easy (just drop the dead code and add something new), and I bumped
up coverage a wee bit in the process. (It'd be easy to keep going, but I hate
mocks, and think most of these tests have little value.)

Closes #90 (via different means).