Fix a bad bounds check for downloads.
Previously, if we created a download with `auto_transfer=True` with the full
download smaller than the default `chunksize`, and then called
`StreamInChunks`, we'd end up downloading the full content twice. This was due
to a bad bounds check in setting a range header -- we'd end up with a header
like `Range: bytes=100-99`, and the server would ignore the start value,
giving us all the bytes again.

This adds a fix and a test.