Fix GCE creds serialization w/ oauth2client >= 4.0 (#195)
Fix GCE creds serialization w/ oauth2client >= 4.0

When updating gsutil to use oauth2client 4.1.2 and apitools 0.5.19,
testing on GCE instances showed that GceAssertionCredentials were not
being written to our credential storage file. Debugging showed that this
was because of a KeyError when looking for the 'scope' attribute in a
serialized credential (no longer included as of oauth2client 3.0). This
change maintains backward compatibility for oauth2client versions that
used the 'scope' attribute while also working for current versions that
don't include 'scope' when constructing a GceAssertionCredentials
object.