Add a function for detecting unrecognized fields.
If this decoding library is used for parsing user input, an unrecognized
field is very frequently a typo/error and the library should offer an
interface to treat them as such. Without any interface for this, library
users must write their own error-detecting/handling routines to parse
the obscure apitools decoding artifacts.

Given that apitools is the one generating this structure, it makes much
more sense for this error-detecting logic to be provided by apitools
directly.