exclude samples & tests from install
There's no value in installing the sample modules (plus they pollute
the global samples.* namespace), nor in installing the unittests.

For the samples, we set the include option with find_packages so the
code only searches the top level apitools dir.

For the tests, because they live alongside modules vs a completely
sep tree (which is fine), we have to add a series of globs to the
exclude_package_data to filter them out.