Pass empty strings instead of None to urllib.parse.urlunsplit. (#221)
* Pass empty strings instead of None to urllib.parse.urlunsplit.

This is because urllib_parse.urlunsplit takes a single tuple as an argument
and uses the first index in the tuple to determine if any
encoding needs to be done on the rest of the arguments in the tuple. If None is
passed as the first argument and then a string is passed as a later
argument in the tuple, urlunsplit throws an error.

This also adds a test for this particular code path.